Move intercept fit helper to new module
